WorkerThreadPool - remove busy workers map.
The busyWorkers map was removed to decouple the pool from the tasks that the workers execute. To make sure that unmanaged workers are not accidentally added to the pool, a Set of references to the workers is created on WorkerThreadPool's creation, allowing for O(1) checks in the returnWorker() method. When shutting down the worker pool, the set of workers is iterated in order to interrupt each thread.