Switch running flag to isShutdown in WorkerThreadPool.
Maintaining a running flag doesn't make sense in the WorkerThreadPool, as it's not a thread - it only manages threads. As we can only shut it down, it makes more sense to maintain an isShutdown flag and check for it instead. The only place where this flag will change is in the shutdown() method, where it is set to true. All operations on the pool will throw IllegalStateException after shutting it down.